refactor: move settings dialog into dev source code
