+++ app/src/dev/java/xyz/lbres/trickcalculator/settings/SettingsDialog.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.trickcalculator.ui.settings.components
+package xyz.lbres.trickcalculator.settings
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -13,6 +13,10 @@ import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.ViewModelProvider
 import xyz.lbres.trickcalculator.R
 import xyz.lbres.trickcalculator.databinding.DialogSettingsBinding
+import xyz.lbres.trickcalculator.ui.settings.components.SettingsUI
+import xyz.lbres.trickcalculator.ui.settings.components.closePreviousFragment
+import xyz.lbres.trickcalculator.ui.settings.components.initUi
+import xyz.lbres.trickcalculator.ui.settings.components.saveToViewModel
 import xyz.lbres.trickcalculator.ui.shared.SharedViewModel
 
 /**
